requirements.txt
opencv-python==4.6.0.66
matplotlib==3.7.4
numpy==1.24.3
torch==2.1.0+cu121
yaml==0.2.5
pyyaml==6.0.1
pathlib==1.0.1
pillow==10.1.0
scipy==1.10.1
tqdm==4.66.1
torchvision==0.16.0+cu121
pandas==2.0.3
seaborn==0.13.0
detectron2==0.5
requests==2.31.0
thop==0.1.1-2209072238
xlwt==1.3.0
xlrd==2.0.1
xlutils==2.0.0
scikit-image==0.19.2
numba==0.58.1
openpyxl==3.0.9
scikit-learn==1.0.2
lxml==4.9.4